Avoid storing sensitive data in the config file
We do this by specifying for every ConfigValue if it is sensitive information
and storing those sensitive values in a separate config object (named
sensitive_config). Only the regular 'config' object is written back to the
file.

In practice this has the effect that you will be asked for sensitive
information (e.g. JIRA password) on every application start.