Add --save-passwords option to force saving sensitive data to config file
This option causes the config and sensitive_config to be combined into one
config that is then written to the config file.

This option can be used to force the old behaviour of writing all information
to the config file.

Once the data is written to the config file the option no longer needs to be
specified because on starting the application the user will not be queried for
information.

Basically the behaviour is as follows:

Scenario 1
----------

Application is started for the first time WITHOUT --save-passwords and no
config file exists.

The user is queried for all data and only the non-sensitive information is
stored in the config file. On subsequent application starts the user will
always be queried for the sensitive data. The user may at any point start the
application with the --save-passwords option and all sensitive data will be
written to the config file as well.

This is the default new behaviour. In practice it means you will be queried on
every application start for sensitive information (e.g. JIRA password).

Scenario 2
----------

Application is started for the first time WITH --save-passwords and no config
file exists.

The user is queried for all data and all this data is stored to the config
file. On subsequent application starts the user will no longer be queried for
any information.

This corresponds to the old behaviour.

Scenario 3
----------

Application is started and the config file already contains all data. This also
applies to the case of updating from an old version of this package to a new
version.

The user will not be queried for any information and everything will be written
back to the config file as it is.

If any data is missing from the config file the user will be queried for that
data. Whether that data is then written back to the config file depends on the
--save-passwords option and on whether the data is sensitive or not (see
scenarios 1 and 2).